+++ Project.toml
@@ -7,14 +7,29 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 IntervalBoxes = "43d83c95-ebbb-40ec-8188-24586a1458ed"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-ReachabilityBase = "379f33d0-9447-4353-bd03-d664070e549f"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
+AffineArithmetic = "0.3"
 ForwardDiff = "0.10, 1"
 IntervalArithmetic = "0.22.12 - 0.23, 1"
 IntervalBoxes = "0.2 - 0.3"
+IntervalOptimisation = "0.5"
 LinearAlgebra = "<0.0.1, 1.10"
-ReachabilityBase = "0.1.1 - 0.3"
-Requires = "0.5, 1"
+MultivariatePolynomials = "0.3 - 0.5"
+SumOfSquares = "0.3.6 - 0.7"
+TaylorModels = "0.9 - 0.10"
 julia = "1.10"
+
+[extensions]
+AffineArithmeticExt = "AffineArithmetic"
+IntervalOptimisationExt = "IntervalOptimisation"
+MultivariatePolynomialsExt = "MultivariatePolynomials"
+SumOfSquaresExt = ["MultivariatePolynomials", "SumOfSquares"]
+TaylorModelsExt = "TaylorModels"
+
+[weakdeps]
+AffineArithmetic = "2e89c364-fad6-56cb-99bd-ebadcd2cf8d2"
+IntervalOptimisation = "c7c68f13-a4a2-5b9a-b424-07d005f8d9d2"
+MultivariatePolynomials = "102ac46a-7ee4-5c85-9060-abc95bfdeaa3"
+SumOfSquares = "4b9e565b-77fc-50a5-a571-1244f986bda1"
+TaylorModels = "314ce334-5f6e-57ae-acf6-00b6e903104a"
