+++ docs/Project.toml
@@ -11,13 +11,9 @@ SumOfSquares = "4b9e565b-77fc-50a5-a571-1244f986bda1"
 TaylorModels = "314ce334-5f6e-57ae-acf6-00b6e903104a"
 
 [compat]
-AffineArithmetic = "0.3"
 DisplayAs = "0.1"
 Documenter = "1"
 DocumenterCitations = "1.3"
 DynamicPolynomials = "0.3 - 0.6"
-IntervalOptimisation = "0.5"
 Plots = "1"
 SDPA = "0.2 - 0.6"
-SumOfSquares = "0.3.6 - 0.7"
-TaylorModels = "0.9 - 0.10"
