+++ ext/SumOfSquaresExt.jl
@@ -1,9 +1,20 @@
-using .SumOfSquares: SOSModel, @set, @variable, @constraint, @objective,
-                     optimize!, objective_value
-import .SumOfSquares: SemialgebraicSets  # `@set` assumes that `SemialgebraicSets` is defined
+module SumOfSquaresExt
 
-function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike,
-                      dom::Interval_or_IntervalVector_or_IntervalBox; kwargs...)
+import RangeEnclosures
+using RangeEnclosures: Interval_or_IntervalVector_or_IntervalBox, SumOfSquaresEnclosure
+
+using IntervalArithmetic: inf, interval, sup
+
+import MultivariatePolynomials
+using MultivariatePolynomials: AbstractPolynomialLike, variables
+
+import SumOfSquares
+using SumOfSquares: SOSModel, @set, @variable, @constraint, @objective, optimize!, objective_value
+import SumOfSquares: SemialgebraicSets  # `@set` assumes that `SemialgebraicSets` is defined
+
+function RangeEnclosures.enclose(p::AbstractPolynomialLike,
+                                 dom::Interval_or_IntervalVector_or_IntervalBox,
+                                 sose::SumOfSquaresEnclosure; kwargs...)
     x = variables(p)
 
     B = reduce(intersect, @set inf(domi) <= xi && xi <= sup(domi) for (xi, domi) in zip(x, dom))
@@ -30,3 +41,5 @@ function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike,
 
     return interval(lower_bound, upper_bound)
 end
+
+end  # module
