+++ test/quality_assurance.jl
@@ -2,8 +2,12 @@ using RangeEnclosures, Test
 import Aqua, ExplicitImports
 
 @testset "ExplicitImports tests" begin
-    ignores_all_explicit_imports_are_public = (:Aff,)
-    ignores_all_qualified_accesses_are_public = (:Fix1, :derivative, :gradient, :where_bisect)
+    ignores_all_explicit_imports_are_public = (:AbstractEnclosureAlgorithm, :Aff,
+                                               :Interval_or_IntervalVector,
+                                               :Interval_or_IntervalVector_or_IntervalBox,
+                                               :IntervalVector_or_IntervalBox)
+    ignores_all_qualified_accesses_are_public = (:Fix1, :derivative, :gradient, :where_bisect,
+                                                 :_default_vector_MSE)
     ignores_no_stale_explicit_imports = (:SemialgebraicSets,)  # false positive due to external macro
     ExplicitImports.test_explicit_imports(RangeEnclosures;
                                           all_explicit_imports_are_public=(ignore=ignores_all_explicit_imports_are_public,),
